+++ src/main/scala/org/openhorizon/exchangeapi/table/Schema.scala
@@ -47,15 +47,15 @@ object SchemaTQ  extends TableQuery(new SchemaTable(_)){
       case 0 => // v1.35.0 - no changes needed to get to time zero
         DBIO.seq()
       case 1 => // v1.37.0
-        DBIO.seq(NodeStatusTQ.schema.createIfNotExists)
+        DBIO.seq(NodeStatusTQ.schema.create)
       case 2 => // v1.38.0
         DBIO.seq(sqlu"ALTER TABLE public.agbots DROP COLUMN IF EXISTS patterns",
-                 AgbotPatternsTQ.schema.createIfNotExists)
+                 AgbotPatternsTQ.schema.create)
       case 3 => // v1.45.0
-        DBIO.seq(ServicesTQ.schema.createIfNotExists)
+        DBIO.seq(ServicesTQ.schema.create)
       case 4 =>
         DBIO.seq(/* WorkloadKeysTQ.schema.create, MicroserviceKeysTQ.schema.create, ServiceKeysTQ.schema.create, */
-                 PatternKeysTQ.schema.createIfNotExists)
+                 PatternKeysTQ.schema.create)
       case 5 => // v1.47.0
         DBIO.seq(sqlu"ALTER TABLE public.patterns ADD COLUMN IF NOT EXISTS services CHARACTER VARYING NOT NULL DEFAULT ''")
       case 6 => // v1.47.0
@@ -76,7 +76,7 @@ object SchemaTQ  extends TableQuery(new SchemaTable(_)){
           actions ++ List(sqlu"ALTER TABLE public.services RENAME COLUMN pkg TO imagestore")
         DBIO.seq(actions: _*) // convert the list of actions to a DBIO seq
       case 9 => // v1.52.0
-        DBIO.seq(ServiceDockAuthsTQ.schema.createIfNotExists)
+        DBIO.seq(ServiceDockAuthsTQ.schema.create)
       case 10 => // v1.56.0
         DBIO.seq(sqlu"ALTER TABLE public.agbotpatterns ADD COLUMN IF NOT EXISTS nodeorgid CHARACTER VARYING NOT NULL DEFAULT ''")
       case 11 => // v1.56.0
@@ -117,18 +117,18 @@ object SchemaTQ  extends TableQuery(new SchemaTable(_)){
       case 17 => // v1.72.0
         DBIO.seq(sqlu"ALTER TABLE public.orgs ADD COLUMN IF NOT EXISTS orgtype CHARACTER VARYING NOT NULL DEFAULT ''")
       case 18 => // v1.77.0
-        DBIO.seq(NodePolicyTQ.schema.createIfNotExists)
+        DBIO.seq(NodePolicyTQ.schema.create)
       case 19 => // v1.77.0
-        DBIO.seq(ServicePolicyTQ.schema.createIfNotExists)
+        DBIO.seq(ServicePolicyTQ.schema.create)
       case 20 => // v1.78.0
-        DBIO.seq(BusinessPoliciesTQ.schema.createIfNotExists)
+        DBIO.seq(BusinessPoliciesTQ.schema.create)
       case 21 => // v1.82.0
         DBIO.seq(sqlu"ALTER TABLE public.services DROP COLUMN IF EXISTS requiredresources",
                  sqlu"DROP TABLE IF EXISTS public.resourcekeys",
                  sqlu"DROP TABLE IF EXISTS public.resourceauths",
                  sqlu"DROP TABLE IF EXISTS public.resources")
       case 22 => // v1.82.0
-        DBIO.seq(AgbotBusinessPolsTQ.schema.createIfNotExists)
+        DBIO.seq(AgbotBusinessPolsTQ.schema.create)
       case 23 => // v1.83.0
         DBIO.seq(sqlu"ALTER TABLE public.nodes ADD COLUMN IF NOT EXISTS arch CHARACTER VARYING NOT NULL DEFAULT ''")
       case 24 => // v1.84.0
@@ -139,11 +139,11 @@ object SchemaTQ  extends TableQuery(new SchemaTable(_)){
       case 26 => // v1.92.0
         DBIO.seq(sqlu"ALTER TABLE public.users ADD COLUMN IF NOT EXISTS updatedBy CHARACTER VARYING NOT NULL DEFAULT ''")
       case 27 => // v1.102.0
-        DBIO.seq(NodeErrorTQ.schema.createIfNotExists)
+        DBIO.seq(NodeErrorTQ.schema.create)
       case 28 => // v1.92.0
         DBIO.seq(sqlu"ALTER TABLE public.nodestatus ADD COLUMN IF NOT EXISTS runningServices CHARACTER VARYING NOT NULL DEFAULT ''")
       case 29 => // v1.122.0
-        DBIO.seq(ResourceChangesTQ.schema.createIfNotExists)
+        DBIO.seq(ResourceChangesTQ.schema.create)
       case 30 => // v2.1.0
         DBIO.seq(sqlu"ALTER TABLE public.nodes ADD COLUMN IF NOT EXISTS heartbeatintervals CHARACTER VARYING NOT NULL DEFAULT ''",
                  sqlu"ALTER TABLE public.orgs ADD COLUMN IF NOT EXISTS heartbeatintervals CHARACTER VARYING NOT NULL DEFAULT ''")
@@ -170,7 +170,7 @@ object SchemaTQ  extends TableQuery(new SchemaTable(_)){
       case 36 => // v2.35.0
         DBIO.seq(sqlu"ALTER TABLE public.nodes ALTER COLUMN IF EXISTS lastheartbeat DROP NOT NULL")
       case 37 =>
-        DBIO.seq(SearchOffsetPolicyTQ.schema.createIfNotExists)
+        DBIO.seq(SearchOffsetPolicyTQ.schema.create)
       case 38 => // v2.42.0
         DBIO.seq(sqlu"ALTER TABLE public.users ADD COLUMN IF NOT EXISTS hubadmin boolean DEFAULT FALSE",
                  sqlu"ALTER TABLE public.orgs ADD COLUMN IF NOT EXISTS limits CHARACTER VARYING NOT NULL DEFAULT ''")
@@ -185,7 +185,7 @@ object SchemaTQ  extends TableQuery(new SchemaTable(_)){
         DBIO.seq(sqlu"ALTER TABLE public.patterns ADD COLUMN IF NOT EXITS secretbinding CHARACTER VARYING NOT NULL DEFAULT ''",
                  sqlu"ALTER TABLE public.businesspolicies ADD COLUMN IF NOT EXISTS secretbinding CHARACTER VARYING NOT NULL DEFAULT ''")
       case 42 => // v2.89.0
-        DBIO.seq(ManagementPoliciesTQ.schema.createIfNotExists)
+        DBIO.seq(ManagementPoliciesTQ.schema.create)
       case 43 => // v2.57.0
         DBIO.seq(sqlu"ALTER TABLE public.nodepolicies ADD COLUMN IF NOT EXISTS deployment CHARACTER VARYING NOT NULL DEFAULT ''",
                  sqlu"ALTER TABLE public.nodepolicies ADD COLUMN IF NOT EXISTS management CHARACTER VARYING NOT NULL DEFAULT ''",
@@ -198,13 +198,12 @@ object SchemaTQ  extends TableQuery(new SchemaTable(_)){
                  sqlu"ALTER TABLE public.managementpolicies ADD COLUMN IF NOT EXISTS startwindow BIGINT NOT NULL DEFAULT 0")
       // NODE: IF ADDING A TABLE, DO NOT FORGET TO ALSO ADD IT TO ExchangeApiTables.initDB and dropDB
       case 45 => // v2.95.0
-        DBIO.seq(sqlu"DROP INDEX IF EXISTS pk_management_policy_status_node",
-                 NodeMgmtPolStatuses.schema.createIfNotExists)
+        DBIO.seq(NodeMgmtPolStatuses.schema.create)
       case 46 => // v2.96.0
-        DBIO.seq(AgentCertificateVersionsTQ.schema.createIfNotExists,
-                 AgentConfigurationVersionsTQ.schema.createIfNotExists,
-                 AgentSoftwareVersionsTQ.schema.createIfNotExists,
-                 AgentVersionsChangedTQ.schema.createIfNotExists)
+        DBIO.seq(AgentCertificateVersionsTQ.schema.create,
+                 AgentConfigurationVersionsTQ.schema.create,
+                 AgentSoftwareVersionsTQ.schema.create,
+                 AgentVersionsChangedTQ.schema.create)
       case 47 => // v2.100.0
         DBIO.seq(sqlu"ALTER TABLE public.management_policy_status_node ALTER COLUMN time_start_actual DROP NOT NULL",
                  sqlu"ALTER TABLE public.management_policy_status_node ALTER COLUMN version_certificate DROP NOT NULL",
@@ -221,18 +220,22 @@ object SchemaTQ  extends TableQuery(new SchemaTable(_)){
                  sqlu"CREATE UNIQUE INDEX IF NOT EXISTS idx_avconfig_priority ON public.agent_version_configuration(organization, priority);",
                  sqlu"""CREATE UNIQUE INDEX IF NOT EXISTS idx_avsoft_priority ON public.agent_version_software("version", priority);""")
       case 49 =>
-        DBIO.seq(NodeGroupTQ.schema.createIfNotExists,
-                 NodeGroupAssignmentTQ.schema.createIfNotExists)
+        DBIO.seq(NodeGroupTQ.schema.create,
+                 NodeGroupAssignmentTQ.schema.create)
       case 50 => // v2.108.0
         DBIO.seq(sqlu"ALTER TABLE public.node_group ALTER COLUMN description DROP NOT NULL;")
       case 51 => // v2.109.0
         DBIO.seq(sqlu"ALTER TABLE public.node_group ADD COLUMN IF NOT EXISTS admin BOOL NOT NULL DEFAULT FALSE;")
+      case 52 => // v2.111.0
+        DBIO.seq(sqlu"ALTER TABLE public.businesspolicies ADD COLUMN IF NOT EXISTS cluster_namespace varchar NULL;",
+                           sqlu"ALTER TABLE public.nodes ADD COLUMN IF NOT EXISTS cluster_namespace varchar NULL;",
+                           sqlu"ALTER TABLE public.patterns ADD COLUMN IF NOT EXISTS cluster_namespace varchar NULL;")
       case other => // should never get here
         logger.error("getUpgradeSchemaStep was given invalid step "+other); DBIO.seq()
     }
   }
 
-  val latestSchemaVersion: Int = 51    // NOTE: THIS MUST BE CHANGED WHEN YOU ADD TO getUpgradeSchemaStep() above
+  val latestSchemaVersion: Int = 52    // NOTE: THIS MUST BE CHANGED WHEN YOU ADD TO getUpgradeSchemaStep() above
   val latestSchemaDescription: String = ""
   // Note: if you need to manually set the schema number in the db lower: update schema set schemaversion = 12 where id = 0;