+++ src/test/scala/org/openhorizon/exchangeapi/NodesSuite.scala
@@ -332,7 +332,7 @@ class NodesSuite extends AnyFunSuite {
   test("POST /orgs/"+orgid+"/business/policies/"+businessPolicySdr+" - add "+businessPolicySdr+" as user") {
     val input = PostPutBusinessPolicyRequest(businessPolicySdr, Some("desc"),
       BService(SDRSPEC_URL, orgid, "*", List(BServiceVersions(svcversion, None, None)), None ),
-      None, None, Some(List(OneProperty("purpose",None,"location"))), Some(List("a == b"))
+      None, None, Some(List(OneProperty("purpose",None,"location"))), Some(List("a == b")), None
     )
     val response = Http(URL+"/business/policies/"+businessPolicySdr).postData(write(input)).method("post").headers(CONTENT).headers(ACCEPT).headers(USERAUTH).asString
     info("code: "+response.code+", response.body: "+response.body)
@@ -342,7 +342,7 @@ class NodesSuite extends AnyFunSuite {
   test("POST /orgs/"+orgid+"/business/policies/"+businessPolicySdr2+" - add "+businessPolicySdr2+" as user") {
     val input = PostPutBusinessPolicyRequest(businessPolicySdr2, Some("desc"),
       BService(SDRSPEC_URL, orgid, "", List(BServiceVersions(svcversion, None, None)), None ),
-      None, None, Some(List(OneProperty("purpose",None,"location"))), Some(List("a == b"))
+      None, None, Some(List(OneProperty("purpose",None,"location"))), Some(List("a == b")), None
     )
     val response = Http(URL+"/business/policies/"+businessPolicySdr2).postData(write(input)).method("post").headers(CONTENT).headers(ACCEPT).headers(USERAUTH).asString
     info("code: "+response.code+", response.body: "+response.body)
@@ -358,7 +358,7 @@ class NodesSuite extends AnyFunSuite {
   test("POST /orgs/"+orgid+"/business/policies/"+businessPolicyNS+" - add "+businessPolicyNS+" as user") {
     val input = PostPutBusinessPolicyRequest(businessPolicyNS, Some("desc"),
       BService(NETSPEEDSPEC_URL, orgid, "amd64", List(BServiceVersions(svcversion, None, None)), None ),
-      None, None, Some(List(OneProperty("purpose",None,"location"))), Some(List("a == b"))
+      None, None, Some(List(OneProperty("purpose",None,"location"))), Some(List("a == b")), None
     )
     val response = Http(URL+"/business/policies/"+businessPolicyNS).postData(write(input)).method("post").headers(CONTENT).headers(ACCEPT).headers(USERAUTH).asString
     info("code: "+response.code+", response.body: "+response.body)